# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  unit: string;
  farmer: string;
  farmerId?: string;
  location: string;
  image: string;
  rating: number;
  reviews: number;
  availableQuantity: number;
  verified: boolean;
  harvest?: string;
  description?: string;
}

export interface Farmer {
  id: string;
  name: string;
  farmName: string;
  location: string;
  image: string;
  rating: number;
  completedOrders: number;
  yearsFarming: number;
  verified: boolean;
  specialty?: string;
}

export interface HarvestInfo {
  harvestDate: string;
  available: string;
  freshness: string;
  nextHarvest: string;
}

export interface Review {
  id: string;
  reviewer: string;
  rating: number;
  date: string;
  text: string;
  verifiedPurchase: boolean;
}

export type ProductDetail = Product & {
  images: string[];
  minimumOrder: number;
  delivery: string;
  farmerProfile: Farmer;
  harvestInfo: HarvestInfo;
  reviewItems: Review[];
};

export type FarmerDetail = Farmer & {
  about?: string;
  farmSize?: string;
  deliveryAreas?: string[];
  products?: Product[];
};

export interface MarketPrice {
  product: string;
  currentPrice: number;
  previousPrice: number;
  change: number;
  market: string;
  date: string;
}

export interface MarketplaceSummary {
  farms: number;
  products: number;
  counties: number;
  orders: number;
  averageRating: number;
}

export type GetProductsParams = {
search?: string;
category?: string;
location?: string;
sort?: GetProductsSort;
};

export type GetProductsSort = typeof GetProductsSort[keyof typeof GetProductsSort];


export const GetProductsSort = {
  recommended: 'recommended',
  'price-low': 'price-low',
  'price-high': 'price-high',
  newest: 'newest',
  popular: 'popular',
} as const;