# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: marketplace
    description: Marketplace browsing and market intelligence
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: getProducts
      tags: [marketplace]
      summary: List marketplace products
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: location
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [recommended, price-low, price-high, newest, popular] }
      responses:
        "200":
          description: Marketplace products
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [marketplace]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Product detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductDetail" }
        "404":
          description: Product not found
  /farmers:
    get:
      operationId: getFarmers
      tags: [marketplace]
      summary: List featured farmers
      responses:
        "200":
          description: Farmers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Farmer" }
  /farmers/{id}:
    get:
      operationId: getFarmer
      tags: [marketplace]
      summary: Get a farmer profile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Farmer profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FarmerDetail" }
  /market-prices:
    get:
      operationId: getMarketPrices
      tags: [marketplace]
      summary: Get sample market price intelligence
      responses:
        "200":
          description: Sample market prices
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/MarketPrice" }
  /marketplace/summary:
    get:
      operationId: getMarketplaceSummary
      tags: [marketplace]
      summary: Get marketplace summary metrics
      responses:
        "200":
          description: Marketplace summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MarketplaceSummary" }
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Product:
      type: object
      required: [id, name, category, price, unit, farmer, location, image, rating, reviews, availableQuantity, verified]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        price: { type: number }
        unit: { type: string }
        farmer: { type: string }
        farmerId: { type: string }
        location: { type: string }
        image: { type: string }
        rating: { type: number }
        reviews: { type: number }
        availableQuantity: { type: number }
        verified: { type: boolean }
        harvest: { type: string }
        description: { type: string }
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/Product"
        - type: object
          required: [images, minimumOrder, delivery, farmerProfile, harvestInfo, reviewItems]
          properties:
            images:
              type: array
              items: { type: string }
            minimumOrder: { type: number }
            delivery: { type: string }
            farmerProfile: { $ref: "#/components/schemas/Farmer" }
            harvestInfo: { $ref: "#/components/schemas/HarvestInfo" }
            reviewItems:
              type: array
              items: { $ref: "#/components/schemas/Review" }
    Farmer:
      type: object
      required: [id, name, farmName, location, image, rating, completedOrders, yearsFarming, verified]
      properties:
        id: { type: string }
        name: { type: string }
        farmName: { type: string }
        location: { type: string }
        image: { type: string }
        rating: { type: number }
        completedOrders: { type: number }
        yearsFarming: { type: number }
        verified: { type: boolean }
        specialty: { type: string }
    FarmerDetail:
      allOf:
        - $ref: "#/components/schemas/Farmer"
        - type: object
          properties:
            about: { type: string }
            farmSize: { type: string }
            deliveryAreas:
              type: array
              items: { type: string }
            products:
              type: array
              items: { $ref: "#/components/schemas/Product" }
    HarvestInfo:
      type: object
      required: [harvestDate, available, freshness, nextHarvest]
      properties:
        harvestDate: { type: string }
        available: { type: string }
        freshness: { type: string }
        nextHarvest: { type: string }
    Review:
      type: object
      required: [id, reviewer, rating, date, text, verifiedPurchase]
      properties:
        id: { type: string }
        reviewer: { type: string }
        rating: { type: number }
        date: { type: string }
        text: { type: string }
        verifiedPurchase: { type: boolean }
    MarketPrice:
      type: object
      required: [product, currentPrice, previousPrice, change, market, date]
      properties:
        product: { type: string }
        currentPrice: { type: number }
        previousPrice: { type: number }
        change: { type: number }
        market: { type: string }
        date: { type: string }
    MarketplaceSummary:
      type: object
      required: [farms, products, counties, orders, averageRating]
      properties:
        farms: { type: number }
        products: { type: number }
        counties: { type: number }
        orders: { type: number }
        averageRating: { type: number }