# lib/api-zod/src/generated/types/farmer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Farmer {
  id: string;
  name: string;
  farmName: string;
  location: string;
  image: string;
  rating: number;
  completedOrders: number;
  yearsFarming: number;
  verified: boolean;
  specialty?: string;
}