# lib/api-zod/src/generated/types/farmerDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Farmer } from './farmer';
import type { Product } from './product';

export type FarmerDetail = Farmer & {
  about?: string;
  farmSize?: string;
  deliveryAreas?: string[];
  products?: Product[];
};