# lib/api-zod/src/generated/types/harvestInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface HarvestInfo {
  harvestDate: string;
  available: string;
  freshness: string;
  nextHarvest: string;
}