# lib/api-zod/src/generated/types/marketPrice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketPrice {
  product: string;
  currentPrice: number;
  previousPrice: number;
  change: number;
  market: string;
  date: string;
}