# lib/api-zod/src/generated/types/marketplaceSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MarketplaceSummary {
  farms: number;
  products: number;
  counties: number;
  orders: number;
  averageRating: number;
}