# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  category: string;
  price: number;
  unit: string;
  farmer: string;
  farmerId?: string;
  location: string;
  image: string;
  rating: number;
  reviews: number;
  availableQuantity: number;
  verified: boolean;
  harvest?: string;
  description?: string;
}