# lib/api-zod/src/generated/types/productDetail.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Farmer } from './farmer';
import type { HarvestInfo } from './harvestInfo';
import type { Product } from './product';
import type { Review } from './review';

export type ProductDetail = Product & {
  images: string[];
  minimumOrder: number;
  delivery: string;
  farmerProfile: Farmer;
  harvestInfo: HarvestInfo;
  reviewItems: Review[];
};