# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  reviewer: string;
  rating: number;
  date: string;
  text: string;
  verifiedPurchase: boolean;
}